% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITAP_common_docs.R
\name{args}
\alias{args}
\title{Common arguments for \code{\link[=flow_mapper]{flow_mapper()}}, \code{\link[=form_mapper]{form_mapper()}}, \code{\link[=facet_mapper]{facet_mapper()}}, a
nd \code{\link[=wepp_mapper]{wepp_mapper()}}}
\arguments{
\item{grid}{Numeric. Grid size in m of the input DEM file}

\item{min_x}{Numeric. Starting x coordinate (in meters)}

\item{min_y}{Numeric. Starting y coordinate (in meters)}

\item{resume}{Character. From which stage should the run be resumed? (see}

\item{clean}{Logical. Remove all output files from previous runs in this
folder?}

\item{report}{Logical. Create html report of results?}

\item{log}{Logical. Create log file recording progress?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}

\item{debug}{Logical. If TRUE, output files contain intermediate columns
useful for debugging (e.g., 'buffer', 'seqno_buffer', etc.) Default FALSE.}
}
\description{
Common arguments for \code{\link[=flow_mapper]{flow_mapper()}}, \code{\link[=form_mapper]{form_mapper()}}, \code{\link[=facet_mapper]{facet_mapper()}}, a
nd \code{\link[=wepp_mapper]{wepp_mapper()}}
}
\keyword{internal}
