% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITAP.R
\name{.onLoad}
\alias{.onLoad}
\title{Dealing with CRAN Notes due to Non-standard evaluation}
\usage{
.onLoad(libname = find.package("LITAP"), pkgname = "LITAP")
}
\description{
Dealing with CRAN Notes due to Non-standard evaluation
}
\keyword{internal}
