% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_mapper.R
\name{facet_mapper}
\alias{facet_mapper}
\title{Calculate landform facets using LSM logic}
\usage{
facet_mapper(folder, arule, crule, out_format = "rds", resume = NULL,
  end = NULL, log = TRUE, report = TRUE, verbose = FALSE,
  quiet = FALSE)
}
\arguments{
\item{folder}{Character. Location of \code{\link[=flow_mapper]{flow_mapper()}} output}

\item{resume}{Character. From which stage should the run be resumed? (see
Details below)}

\item{end}{Character. If ending a run after a particular step, which step
(see Details below)}

\item{log}{Logical. Create log file recording progress?}

\item{report}{Logical. Create html report of results?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}
}
\description{
This function takes backup data frame output from \code{\link[=flow_mapper]{flow_mapper()}} and
\code{\link[=form_mapper]{form_mapper()}} calculates fuzzy attributes (among other metrics).  Based on
FacetMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions.
}
\details{
For resuming or ending a run, \code{resume} or \code{end} must be
one of the following:
\itemize{
\item facet
}
}
