% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_mapper.R
\name{flow_mapper}
\alias{flow_mapper}
\title{Map flow through the landscape}
\usage{
flow_mapper(
  file,
  nrow,
  ncol,
  grid = NULL,
  min_x = 1,
  min_y = 1,
  missing_value = -9999,
  max_area = 10,
  max_depth = 0.5,
  out_folder = NULL,
  out_format = "rds",
  clean = FALSE,
  clim = NULL,
  rlim = NULL,
  resume = NULL,
  log = TRUE,
  report = TRUE,
  pitr_method = "landmapr",
  verbose = FALSE,
  quiet = FALSE,
  debug = FALSE
)
}
\arguments{
\item{file}{Character. Elevation file (see \code{\link{load_file}}) for
supported file types.}

\item{nrow}{Numeric. Number of rows in dem file (required for dbf files with
a single column, but can be automatically assessed from files with x and y
coordinates.}

\item{ncol}{Numeric. Number of columns in dem file (required for dbf files
with a single column, but can be automatically assessed from files with x
and y coordinates.}

\item{grid}{Numeric. Grid size in m of the input DEM file}

\item{min_x}{Numeric. Starting x coordinate (in meters)}

\item{min_y}{Numeric. Starting y coordinate (in meters)}

\item{missing_value}{Numeric/Character. Symbols which define missing data}

\item{max_area}{Numeric. Largest area of pits to be removed during initial
pit removal}

\item{max_depth}{Numeric. Largest depth of pits to be removed during initial
pit removal}

\item{out_folder}{Character. Folder in which to store output files. Defaults
to folder in the same location and with the same name as the dem file}

\item{out_format}{Character. What format should the data be output as? "rds"
for R data format (default), "csv" for Comma-separated values. This format
is used for all subsequent functions (i.e. \code{form_mapper()},
\code{facet_mapper()} and \code{wepp_mapper()}.}

\item{clean}{Logical. Remove all backup files and output files from previous
runs in this folder?}

\item{clim}{Numeric vector. Column limits if specifying a subset of the dem}

\item{rlim}{Numeric vector. Row limits if specifying a subset of the dem}

\item{resume}{Character. From which stage should the run be resumed? (see}

\item{log}{Logical. Create log file recording progress?}

\item{report}{Logical. Create html report of results?}

\item{pitr_method}{Character. Either "landmapr" or "litap" to define whether
to use the LandMapR method for finding a pour point during the pit removal stage
(default) or LITAP.
The LITAP method defines the pourpoint as the lowest neighbour. The LandMapR
defines it as the last \emph{possible} pour point among neighbouring cells (not
necessarily the lowest neighbouring cell).}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}

\item{debug}{Logical. If TRUE, output files contain intermediate columns
useful for debugging (e.g., 'buffer', 'seqno_buffer', etc.) Default FALSE.}
}
\description{
Run an elevation file through all functions to calculate watershed flow and
fill patterns. Based on FlowMapR by R. A. (Bob) MacMillan, LandMapper
Environmental Solutions.
}
\details{
For information regarding loading other file types see
\code{\link{load_file}}.

For resuming a run, \code{resume} must be one of
the following:

\enumerate{
\item \code{directions} (Calculating Directions)
\item \code{watersheds} (Calculating Watersheds)
\item \code{local} (Initial Pit Removal)
\item \code{pond} (Calculating Pond Shed Statistics - Second Pit Removal)
\item \code{fill} (Calculating Fill Shed Statistics - Third Pit Removal)
\item \code{slope} (Slope Gradient and Curvature values)
\item \code{idirections} (Calculating Directions on Inverted DEM)
\item \code{iwatersheds} (Calculating Inverted Watersheds)
\item \code{inverted} (Inverted Pit Removal)
\item \code{report} (Create the final report)
}
}
\examples{
# Basic Run
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 90, ncol = 90, grid = 1)

# Specify parameters for initial pit removal
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
            out_folder = "./testELEV/", nrow = 90, ncol = 90, grid = 1,
            max_area = 5, max_depth = 2)

# Clean up (remove created folder and output)
unlink("./testELEV/", recursive = TRUE)

}
