% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_plot.R
\name{flow_plot}
\alias{flow_plot}
\title{Create plots of water flow}
\usage{
flow_plot(db, type = "relief", dir = FALSE, seqno = FALSE,
  highlight = FALSE, shed = FALSE, shed_type = "shedno", pits = FALSE,
  upslope_threshold = NULL, cells = NULL, clim = NULL, rlim = NULL,
  stats = NULL, missing = NA)
}
\arguments{
\item{db}{Data frame. Cell by cell data on the elevation of the watershed.
Output by LITAP's \code{complete_run()} function.}

\item{type}{Character. Either relief or elevation. Defaults to relief.}

\item{dir}{Logical. Include flow directions?}

\item{seqno}{Logical. Include cell numbering?}

\item{highlight}{Logical. Highlight selected cells?}

\item{shed}{Logical. Show watersheds?}

\item{shed_type}{Character. Which type of watershed, must be included as a
column in the data frame. Can be one of 'initial', 'local', 'pond', or
'fill'.}

\item{pits}{Logical. Show watershed pits (lowest point)}

\item{upslope_threshold}{Numeric. If dir = TRUE, only show flow directions
for cells with >= this many cells which drain to it.}

\item{cells}{Vector. Which cells to show}

\item{clim}{Numeric vector. Column limits in format c(0, 100)}

\item{rlim}{Numeric vector. Row limits in format c(0, 100)}

\item{stats}{Data frame. Data frame of watershed stats to highlight pour
points.}

\item{missing}{Character. What is the value of missing data? Defaults to NA}
}
\description{
Plots water flow and watersheds. See the flow_plot article/vignette for
examples.
}
