% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_mapper.R
\name{form_mapper}
\alias{form_mapper}
\title{Map form and relief of the landscape}
\usage{
form_mapper(folder, grid, resume = NULL, end = NULL, log = TRUE,
  report = TRUE, verbose = FALSE, quiet = FALSE)
}
\arguments{
\item{folder}{Character. Location of [flow_mapper()] output}

\item{grid}{Numeric. Grid size for the original dem}

\item{resume}{Character. From which stage should the run be resumed? (see
Details below)}

\item{end}{Character. If ending a run after a particular step, which step
(see Details below)}

\item{log}{Logical. Create log file recording progress?}

\item{report}{Logical. Create html report of results?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}
}
\description{
This function takes backup data frame output from [flow_mapper()] and
calculates form, wetness indices, reflief and stream/crest lengths (among
other metrics).  Based on FormMapR by R. A. (Bob) MacMillan, LandMapper
Environmental Solutions.
}
\details{
Runs can be resumed at the "relief" step or the "length" step.

  For resuming or ending a run, \code{resume} or \code{end} must be one of
  the following:

  1. `relief` (Calculating Relief Derivitives)
  2. `length` (Calculating Slope Length)
}
\examples{

\dontrun{
form_mapper(folder = "my_runs/021/", grid = 5)
}

}
