% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITAP-package.R
\docType{package}
\name{LITAP-package}
\alias{LITAP}
\alias{LITAP-package}
\title{LITAP: Landscape Integrated Terrain Analysis Package}
\description{
Terrain analysis and landscape and hydrology models built on terrain attributes. A major component of LITAP is founded on R. A. (Bob) MacMillan's LandMapR suite of programs for flow topology and landform segmentation analyses with extended new parameters and methodologies, as well as with new calculations and uses of directional terrain attributes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRDC-SHL/LITAP}
}

}
\author{
\strong{Maintainer}: Steffi LaZerte \email{steffi@steffi.ca}

Authors:
\itemize{
  \item Sheng Li \email{sheng.li@canada.ca}
}

Other contributors:
\itemize{
  \item Agriculture and Agri-Food Canada [funder]
}

}
\keyword{internal}
