% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITAP_load.R
\name{load_file}
\alias{load_file}
\title{Load and prep elevation data}
\usage{
load_file(
  file,
  nrow = NULL,
  ncol = NULL,
  missing_value = -9999,
  rlim = NULL,
  clim = NULL,
  grid = NULL,
  min_x = 1,
  min_y = 1,
  edge = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file}{Character. The location of the file containing elevation data.
See details for accepted file types}

\item{nrow}{Numeric. Number of rows in dem file (required for dbf files with
a single column, but can be automatically assessed from files with x and y
coordinates.}

\item{ncol}{Numeric. Number of columns in dem file (required for dbf files
with a single column, but can be automatically assessed from files with x
and y coordinates.}

\item{missing_value}{Vector. The number or character string specifying
missing data.}

\item{rlim}{Vector. Two numbers specifying the start and end of a subset of
rows to extract}

\item{clim}{Vector. Two numbers specifying the start and end of a subset of
columns to extract}

\item{grid}{Numeric. Grid size in m of the input DEM file}

\item{min_x}{Numeric. Starting x coordinate (in meters)}

\item{min_y}{Numeric. Starting y coordinate (in meters)}

\item{edge}{Logical. Whether to add an edge (buffer) around the data.}

\item{verbose}{Logical. Output extra progress messages.}
}
\value{
Returns a data frame containing elevation data in a format suitable
for analysis
}
\description{
This function is used by the \code{mapper} functions to load input files and
prepare them for analysis. It can also be used to load input files
independently for plotting with \code{\link{flow_plot}} and/or
trouble-shooting.
}
\details{
All x/y data must be in a format such that greater values indicate
East or North respectively.

This function uses file extensions to guess the file type to be loaded.

\strong{dBase files:}
These files are loaded via the \code{\link[foreign]{read.dbf}} function from
the foreign package. Columns must be named and must have a valid name (case
doesn't not matter). X/Y coordinates are optional and must be named as "x",
"lon", "long", "longitude", or "y", "lat", "latitude". Elevation columns can
be "elev", "elevation", or "z". If no "x" and "y" columns are suppled,
\code{nrow} and \code{ncol} arguments must be supplied. Column names matter,
column order does not. Extra columns, if present, are ignored.
\itemize{
\item dBase files (.dbf)
}

\strong{Grid file types:}
These file types are loaded via the \code{\link[raster]{raster}} function.
\itemize{
\item Surfer grid files (.grd)
\item Esri grid files (binary .adf or ascii .asc)
\item Geo Tiff (.tif)
\item Floating point raster files (.flt) (\strong{Note} the companion header,
.hdr, file must also be present)
}

\strong{Text/Spreadsheet file types:}
Data in these files are all assumed to be arranged in three columns
reflecting x, y, and z dimensions (z = elevation). Column \strong{order} is
important. Column names don't matter, but they should be present.
\itemize{
\item Text files (.txt, .dat, .csv) are loaded via R base
\code{link[utils]{read.table}} function.
\item Excel files (.xls, .xlsx) are loaded via the
\code{\link[readxl]{read_excel}} function.
}
}
