% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITAP_combine_files.R
\name{merge_all}
\alias{merge_all}
\title{Combine flow and form output dems}
\usage{
merge_all(folder, out_format = NULL)
}
\arguments{
\item{folder}{Character. Folder with previous LITAP runs (i.e. where output
of \code{flow_mapper()} etc. are)}

\item{out_format}{Character. Output format (rds or csv) that merged file
should be saved as (if different from the rest; by default uses the format
of the other LITAP output files)}
}
\description{
\code{flow_mapper()} and \code{form_mapper()} each provide output information per cell
of a dem file. This function takes the fill dem from \code{flow_mapper()} as well
as the length and weti dem files from \code{form_mapper()} and merges them
together into a complete dem file with all information. This file is saved
to the project folder.
}
