% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_combine_dems.R
\name{merge_flow_form}
\alias{merge_flow_form}
\title{Combine flow and form output dems}
\usage{
merge_flow_form(folder, out_format = NULL)
}
\arguments{
\item{folder}{Character. Folder with previous LITAP runs (i.e. where output
of \code{flow_mapper()} etc. are)}

\item{out_format}{Character. Output format (rds or csv) that merged file
should be saved as (if different from the rest; by default uses the format
of the other LITAP output files)}
}
\description{
\code{flow_mapper()} and \code{form_mapper()} each provide output information per cell
of a dem file. This function takes the fill dem from \code{flow_mapper()} as well
as the length and weti dem files from \code{form_mapper()} and merges them
together into a complete dem file with all information. This file is saved
to the project folder.
}
\examples{

# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
            grid = 5, nrow = 90, ncol = 90, out_folder = "./testELEV/")

# Then run form_mapper()
form_mapper(folder = "./testELEV/")


# Now merge together
merge_flow_form(folder = "./testELEV")

# If you prefer this file as CSV
merge_flow_form(folder = "./testELEV", out_format = "csv")

# Clean up (remove all output)
unlink("./testELEV/", recursive = TRUE)


}
