% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_slope_gc.R
\name{slope_gc}
\alias{slope_gc}
\title{Calculate slope gradient and curvature}
\usage{
slope_gc(db, grid = 1)
}
\arguments{
\item{db}{Dataframe dem}
}
\value{

}
\description{
Computes row (east/west) and column (north/south) slope gradients and curvatures.
}
\details{
Assume the following cells, and the calculations on elevation for point 5:

c7  c8  c9
c4  c5  c6
c1  c2  c3
\itemize{
\item Row slope gradient towards east (sgre): (c4 - c6) / (2 * grid)
\itemize{
\item Positive slope is downslope (and east-facing)
\item Negative slope is upslope (and west-facing)
}
\item Row slope gradient (sgr): abs(sgre)
\item Column slope gradient towards north (sgcn): (c2 - c8) / (2 * grid)
\itemize{
\item Positive slope is downslope (and north-facing)
\item Negative slope is upslope (and south-facing)
}
\item Column slope gradient (sgc): abs(sgcn)
\item Row slope curvature (scr): (2 * c5 - c4 - c6) / (grid^2)
\item Column slope curvature (scc): (2 * c5 - c2 - c8) / (grid^2)
}

Where missing neighbours, assume same elevation as central point (i.e.
assume an extension of the field).
}
\examples{

d <- slope_gc(test_dem)

library(ggplot2)
flow_plot(d, type = "elevation") +
  geom_point(aes(colour = factor(hill_r_n)))

}
