% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wepp_mapper.R
\name{wepp_mapper}
\alias{wepp_mapper}
\title{Calculate spatial entities required for WEPP}
\usage{
wepp_mapper(folder, out_format = "rds", resume = NULL, end = NULL,
  log = TRUE, report = TRUE, verbose = FALSE, quiet = FALSE)
}
\arguments{
\item{folder}{Character. Location of \code{\link[=flow_mapper]{flow_mapper()}} output}

\item{resume}{Character. From which stage should the run be resumed? (see
Details below)}

\item{end}{Character. If ending a run after a particular step, which step
(see Details below)}

\item{log}{Logical. Create log file recording progress?}

\item{report}{Logical. Create html report of results?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}

\item{grid}{Numeric. Grid size for the original dem}
}
\description{
This function takes backup data frame output from \code{\link[=flow_mapper]{flow_mapper()}} and
\code{\link[=form_mapper]{form_mapper()}} calculates hillslope, channel segments and impoundment
spatial entities required for input in to WEPP (Water Erosion and Prediction
Project) (among other metrics).  Based on FacetMapR by R. A. (Bob) MacMillan,
LandMapper Environmental Solutions.
}
