% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_docs.R
\name{args}
\alias{args}
\title{Common arguments for [flow_mapper()] and [form_mapper()]}
\arguments{
\item{resume}{Character. From which stage should the run be resumed? (see
Details below)}

\item{end}{Character. If ending a run after a particular step, which step
(see Details below)}

\item{report}{Logical. Create html report of results?}

\item{log}{Logical. Create log file recording progress?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}
}
\description{
Common arguments for [flow_mapper()] and [form_mapper()]
}
\keyword{internal}
