% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_run.R
\name{complete_run}
\alias{complete_run}
\title{Complete Run}
\usage{
complete_run(file, nrow, ncol, missing_value = -9999, max_area = 10,
  max_depth = 0.5, folder_out = NULL, clean = FALSE, clim = NULL,
  rlim = NULL, continue = NULL, end = NULL, report = TRUE,
  verbose = FALSE, quiet = FALSE)
}
\arguments{
\item{file}{Character. Elev.dbf dem file}

\item{nrow}{Numeric. Number of rows in dem file}

\item{ncol}{Numeric. Number of columns in dem file}

\item{missing_value}{Numeric/Character. Symbols which define missing data}

\item{max_area}{Numeric. Largest area of pits to be removed during initial
pit removal}

\item{max_depth}{Numeric. Largest depth of pits to be removed during initial
pit removal}

\item{folder_out}{Charater. Folder to store output files. Defaults to
location of dem file if not specified}

\item{clean}{Logical. Remove all backup files and output files from previous
runs?}

\item{clim}{Numeric vector. Column limits if specifying a subset of the dem}

\item{rlim}{Numeric vector. Row limits if specifying a subset of the dem}

\item{continue}{Character. If resuming a run, where to resume (see Details
below)}

\item{end}{Character. If ending a run after a particular step, which step
(see Details below)}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}
}
\description{
Run an elevation dem file through all functions to caculate watershed flow
and fill patterns
}
\details{
For resuming or ending a run, \code{continue} or \code{end} must be
  one of the following:

  \enumerate{
    \item \code{directions} (Calculating Directions)
    \item \code{watersheds} (Calculating Watersheds)
    \item \code{watershed_area} (Calculating Watershed Area
    \item \code{local} (Initial Pit Removal)
    \item \code{pond} (Calculating Pond Shed Statistics - Second Pit Removal)
    \item \code{fill} (Calculating Fill Shed Statistics - Third Pit Removal)
    \item \code{inverted} (Calculating Directions on Inverted DEM)
    \item \code{iwatersheds} (Calculating Inverted Watersheds)
    \item \code{iwatershed_area} (Calculating Watershed Area)
    \item \code{ilocal} (Initial Inverted Pit Removal)
    \item \code{report} (Create the final report)
  }
}
\examples{
\dontrun{
# Basic Run
complete_run(file = "testElev.dbf", nrow = 100, ncol = 100)

# Specify parameters for initial pit removal
complete_run(file = "testElev.dbf", nrow = 100, ncol = 100,
             max_area = 5, max_depth = 2)
}

}
