% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_mapper.R
\name{facet_mapper}
\alias{facet_mapper}
\title{Calculate landform facets using LSM logic}
\usage{
facet_mapper(
  folder,
  arule = NULL,
  crule,
  edge_row = NULL,
  edge_col = NULL,
  procedure = "lsm",
  zone = NULL,
  clean = FALSE,
  resume = NULL,
  log = TRUE,
  verbose = FALSE,
  quiet = FALSE,
  debug = FALSE
)
}
\arguments{
\item{folder}{Character. Location of \code{\link[=flow_mapper]{flow_mapper()}} output}

\item{arule}{Character. Location of ARULE file. If NULL, A Rules are derived
from the dem file (see Details).}

\item{crule}{Character. Location of CRULE file}

\item{edge_row}{Numeric. Number of rows to remove around the edge of the dem
before deriving the A Rules. Default (NULL) results in removing 5\% of the
rows per side (total of 10\%).}

\item{edge_col}{Numeric. Number of cols to remove around the edge of the dem
before deriving the A Rules. Default (NULL) results in removing 5\% of the
cols per side (total of 10\%).}

\item{procedure}{Character. Which procedure to use. One of \code{lsm}
(Original LandMapR program) or \code{bc_pem} (newer BC-PEM Direct-to-Site-Series
program).}

\item{zone}{file. If \code{procedure = "bc_pem"}, zones must either be defined for
each seqno in the form dem file, OR must be provided as an index file
here. With a \code{zone} defined for each \code{seqno}. \code{zone} file can be either dem
(.dem), Excel (.xlsx, .xls), or text (.txt, .csv, .dat)}

\item{clean}{Logical. Remove all output files from previous runs in this
folder?}

\item{resume}{Character. From which stage should the run be resumed? (see}

\item{log}{Logical. Create log file recording progress?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}

\item{debug}{Logical. If TRUE, output files contain intermediate columns
useful for debugging (e.g., 'buffer', 'seqno_buffer', etc.) Default FALSE.}
}
\description{
This function takes backup data frame output from \code{\link[=flow_mapper]{flow_mapper()}} and
\code{\link[=form_mapper]{form_mapper()}} calculates fuzzy attributes (among other metrics).  Based on
FacetMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions.
}
\details{
Based on the technique described in Li et al. 2011, if no \code{arule} file is
provided, the ARULE cutoffs are calculated from the \code{form_mapper()} dem
files. These A Rules are saved as \code{afile_derived.csv} in the \code{folder}
provided. The topographic derivative percentiles are stored to
\code{topographic_derivatives.csv}, also in the \code{folder}.

Procedure \code{lsm} refers to the landform segmentation model (LSM) offered in
the original LandMapR. Procedure \code{bc_pem} refers to calculating variables
required for the British Columbia Predictive Ecosystem Mapping
Direct-to-Site-Series program (BC-PEM DSS).

For resuming  a run, \code{resume} must be
one of the following:
\itemize{
\item attributes
\item classes
}

@references
Sheng Li, David A. Lobb, Brian G. McConkey, R. A. MacMillan, Alan Moulin,
and Walter R. Fraser. 2011. Extracting topographic characteristics of landforms
typical of Canadian agricultural landscapes for agri-environmental modeling.
I. Methodology. Canadian Journal of Soil Science 91(2), 251-266.
\doi{10.4141/CJSS10080}.
}
\examples{

# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 90, ncol = 90, grid = 5)

# And form_mapper()
form_mapper(folder = "./testELEV/")

# Let's use the crule file included in the LITAP package
crule <- system.file("extdata", "crule.dbf", package = "LITAP")
crule

# Run facet_mapper() - Derive A Rules
facet_mapper(folder = "./testELEV/", arule = NULL, crule = crule)

# Derive A Rules, omitting rows and cols from the calculation
facet_mapper(folder = "./testELEV/", arule = NULL, crule = crule,
             edge_row = 3, edge_col = 1)

# Run facet_mapper() - supply A Rules
arule <- system.file("extdata", "arule.dbf", package = "LITAP")
crule <- system.file("extdata", "crule.dbf", package = "LITAP")
facet_mapper(folder = "./testELEV/", arule = arule, crule = crule)

\dontrun{
# Now consider using your own Rule files
facet_mapper(folder = "./testELEV/",
             arule = "./testELEV/my_arule.dbf",
             crule = "./testELEV/my_crule.dbf")

}
# Clean up (remove all output)
unlink("./testELEV/", recursive = TRUE)


}
