% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_mapper.R
\name{flow_mapper}
\alias{flow_mapper}
\title{Map flow through the landscape}
\usage{
flow_mapper(
  file,
  nrow = NULL,
  ncol = NULL,
  missing_value = -9999,
  max_area = 10,
  max_depth = 0.5,
  out_folder = NULL,
  out_format = "rds",
  clean = FALSE,
  clim = NULL,
  rlim = NULL,
  resume = NULL,
  end = NULL,
  log = TRUE,
  report = TRUE,
  verbose = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{file}{Character. Elevation file (see \code{\link{load_file}}) for
supported file types.}

\item{nrow}{Numeric. Number of rows in dem file (required for dbf files with
a single column, but can be automatically assessed from files with x and y
coordinates.}

\item{ncol}{Numeric. Number of columns in dem file (required for dbf files
with a single column, but can be automatically assessed from files with x
and y coordinates.}

\item{missing_value}{Numeric/Character. Symbols which define missing data}

\item{max_area}{Numeric. Largest area of pits to be removed during initial
pit removal}

\item{max_depth}{Numeric. Largest depth of pits to be removed during initial
pit removal}

\item{out_folder}{Character. Folder in which to store output files. Defaults
to folder in the same location and with the same name as the dem file}

\item{clean}{Logical. Remove all backup files and output files from previous
runs in this folder?}

\item{clim}{Numeric vector. Column limits if specifying a subset of the dem}

\item{rlim}{Numeric vector. Row limits if specifying a subset of the dem}

\item{resume}{Character. From which stage should the run be resumed? (see
Details below)}

\item{end}{Character. If ending a run after a particular step, which step
(see Details below)}

\item{log}{Logical. Create log file recording progress?}

\item{report}{Logical. Create html report of results?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}
}
\description{
Run an elevation file through all functions to calculate watershed flow and
fill patterns. Based on FlowMapR by R. A. (Bob) MacMillan, LandMapper
Environmental Solutions.
}
\details{
For information regarding loading other file types see
\code{\link{load_file}}.

For resuming or ending a run, \code{resume} or \code{end} must be one of
the following:

\enumerate{
\item \code{directions} (Calculating Directions)
\item \code{watersheds} (Calculating Watersheds)
\item \code{local} (Initial Pit Removal)
\item \code{pond} (Calculating Pond Shed Statistics - Second Pit Removal)
\item \code{fill} (Calculating Fill Shed Statistics - Third Pit Removal)
\item \code{slope} (Slope Gradient and Curvature values)
\item \code{inverted} (Calculating Directions on Inverted DEM)
\item \code{iwatersheds} (Calculating Inverted Watersheds)
\item \code{ilocal} (Initial Inverted Pit Removal)
\item \code{report} (Create the final report)
}
}
\examples{
# Basic Run
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 150, ncol = 150)

# Specify parameters for initial pit removal
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
            out_folder = "./testELEV/", nrow = 150, ncol = 150,
            max_area = 5, max_depth = 2)

# Clean up (remove created folder and output)
unlink("./testELEV/", recursive = TRUE)

}
