% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_mapper.R
\name{form_mapper}
\alias{form_mapper}
\title{Map form and relief of the landscape}
\usage{
form_mapper(
  folder,
  str_val = 10000,
  ridge_val = 10000,
  resume = NULL,
  log = TRUE,
  clean = FALSE,
  verbose = FALSE,
  quiet = FALSE,
  debug = FALSE
)
}
\arguments{
\item{folder}{Character. Location of \code{\link[=flow_mapper]{flow_mapper()}} output}

\item{str_val}{Numeric. Definition of a stream (number of upslope cells)}

\item{ridge_val}{Numeric. Definition of a ridge (number of downslope cells)}

\item{resume}{Character. From which stage should the run be resumed? (see}

\item{log}{Logical. Create log file recording progress?}

\item{clean}{Logical. Remove all output files from previous runs in this
folder?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}

\item{debug}{Logical. If TRUE, output files contain intermediate columns
useful for debugging (e.g., 'buffer', 'seqno_buffer', etc.) Default FALSE.}
}
\description{
This function takes backup data frame output from \code{\link[=flow_mapper]{flow_mapper()}} and
calculates form, wetness indices, reflief and stream/crest lengths (among
other metrics).  Based on FormMapR by R. A. (Bob) MacMillan, LandMapper
Environmental Solutions.
}
\details{
For resuming a run, \code{resume} must be one of the following:
\enumerate{
\item \code{weti} (Calculating Wetness Indices)
\item \code{relief} (Calculating Relief Derivitives)
\item \code{length} (Calculating Slope Length)
}

Note that some variables have a version 1 and a version 2 (i.e. \code{qweti1} and
\code{qweti2}). These reflect variables calculated (1) area based on number of
cells vs. (2) area based on actual grid cell area values.
}
\examples{

# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 90, ncol = 90, grid = 5)

# Now can run form_mapper()
form_mapper(folder = "./testELEV/")

# Clean up (remove all output)
unlink("./testELEV/", recursive = TRUE)

}
