% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_mapper.R
\name{form_mapper}
\alias{form_mapper}
\title{Map form and relief of the landscape}
\usage{
form_mapper(
  folder,
  grid,
  str_val = 10000,
  ridge_val = 10000,
  out_format = "rds",
  resume = NULL,
  end = NULL,
  log = TRUE,
  clean = FALSE,
  verbose = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{folder}{Character. Location of \code{\link[=flow_mapper]{flow_mapper()}} output}

\item{grid}{Numeric. Grid size for the original dem}

\item{str_val}{Numeric. Definition of a stream (number of upslope cells)}

\item{ridge_val}{Numeric. Definition of a ridge (number of downslope cells)}

\item{out_format}{Character. What format should the data be output as? "csv"
for Comma-separated values, "dbf" for dbf database files, or "rds" for R
data format.}

\item{resume}{Character. From which stage should the run be resumed? (see
Details below)}

\item{end}{Character. If ending a run after a particular step, which step
(see Details below)}

\item{log}{Logical. Create log file recording progress?}

\item{clean}{Logical. Remove all output files from previous runs in this
folder?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}
}
\description{
This function takes backup data frame output from \code{\link[=flow_mapper]{flow_mapper()}} and
calculates form, wetness indices, reflief and stream/crest lengths (among
other metrics).  Based on FormMapR by R. A. (Bob) MacMillan, LandMapper
Environmental Solutions.
}
\details{
For resuming or ending a run, \code{resume} or \code{end} must be
one of the following:
\enumerate{
\item \code{weti} (Calculating Wetness Indices)
\item \code{relief} (Calculating Relief Derivitives)
\item \code{length} (Calculating Slope Length)
}

Note that some variables have a version 1 and a version 2 (i.e. \code{qweti1} and
\code{qweti2}). These reflect variables calculated (1) area based on number of
cells vs. (2) area based on actual grid cell area values.
}
\examples{

# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 150, ncol = 150)

# Now can run form_mapper()
form_mapper(folder = "./testELEV/", grid = 5)

# Clean up (remove all output)
unlink("./testELEV/", recursive = TRUE)

}
