% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_form.R
\name{form_mapper}
\alias{form_mapper}
\title{Calculate form}
\usage{
form_mapper(db, idb, grid, resume = NULL, verbose = TRUE)
}
\arguments{
\item{db}{Data frame. Backup db pond file}

\item{idb}{Data frame. Backup inverted db local file}

\item{grid}{Numeric.    Grid size}

\item{resume}{Character. From which stage should the run be resumed?}

\item{verbose}{Logical. Print progress messages?}
}
\description{
This function takes backup data frame output from [complete_run()] and
calculates form, wetness indices, reflief and stream/crest lengths (among
other metrics). It is based on FormMapR.
}
\details{
Output files are saved the working directory.
}
\examples{

\dontrun{
db <- readr::read_rds("./backup/021_backup_fill.rds")$db
idb <- readr::read_rds("./backup/021_backup_ilocal.rds")$db

form_mapper(db, idb, grid = 5)
}

}
