% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_combine_dems.R
\name{merge_flow_form}
\alias{merge_flow_form}
\title{Combine flow and form output dems}
\usage{
merge_flow_form(folder, format = "rds")
}
\arguments{
\item{folder}{Character. Folder with previous LITAP runs (i.e. where output
of \code{flow_mapper()} etc. are)}

\item{format}{Character. Output format (rds or csv) that files are saved as.}
}
\description{
\code{flow_mapper()} and \code{form_mapper()} each provide output information per cell
of a dem file. This function takes the fill dem from \code{flow_mapper()} as well
as the form, length, and weti dem files from \code{form_mapper()} and merges them
together into a complete dem file with all information.
}
\examples{

# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 150, ncol = 150)

# Then run form_mapper()
form_mapper(folder = "./testELEV/", grid = 5)


# Now merge together
merge_flow_form(folder = "./testELEV")

# Clean up (remove all output)
unlink("./testELEV/", recursive = TRUE)


}
