% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITAP_summary_tables.R
\name{summary_tables}
\alias{summary_tables}
\title{Create summary watersheed tables in Excel}
\usage{
summary_tables(
  folder,
  file = "text.xlsx",
  testing = TRUE,
  min_x = NULL,
  min_y = NULL
)
}
\arguments{
\item{file}{Character. Name of output file (.xlsx)}

\item{min_x}{Numeric. Override the starting x coordinate in the original
data (in meters)}

\item{min_y}{Numeric. Override the starting y coordinate in the original
data (in meters)}
}
\description{
Using output from the flow_mapper(), form_mapper() and facet_maper() runs,
summarizes into various tables in Excel worksheet.
}
\examples{

summary_tables()


}
