% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wepp_mapper.R
\name{wepp_mapper}
\alias{wepp_mapper}
\title{Calculate spatial entities required for WEPP}
\usage{
wepp_mapper(
  folder,
  grid,
  chan_length = 200,
  upslope_threshold = 300,
  clean = FALSE,
  resume = NULL,
  end = NULL,
  log = TRUE,
  verbose = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{folder}{Character. Location of \code{\link[=flow_mapper]{flow_mapper()}} output}

\item{grid}{Numeric. Grid size for the original dem}

\item{chan_length}{Numeric. Channel length maximum length. Used to split
channels into segments}

\item{upslope_threshold}{Numeric. Threshold of upslope cells to define
channel cells. #' Cells with an upslope value larger than this are
considered channel cells}

\item{clean}{Logical. Remove all output files from previous runs in this
folder?}

\item{resume}{Character. From which stage should the run be resumed? (see
Details below)}

\item{end}{Character. If ending a run after a particular step, which step
(see Details below)}

\item{log}{Logical. Create log file recording progress?}

\item{verbose}{Logical. Output extra progress messages.}

\item{quiet}{Logical. Suppress all messages.}
}
\description{
This function takes backup data frame output from \code{\link[=flow_mapper]{flow_mapper()}} and
\code{\link[=form_mapper]{form_mapper()}} calculates hillslope, channel segments and impoundment
spatial entities required for input in to WEPP (Water Erosion and Prediction
Project) (among other metrics).  Based on FacetMapR by R. A. (Bob) MacMillan,
LandMapper Environmental Solutions.
}
\examples{

\dontrun{
# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 90, ncol = 90)

# Now can run wepp_mapper()
wepp_mapper(folder = "./testELEV/", grid = 5)
}

}
